FastAPI2.0/requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-dotenv
bcrypt
passlib
pyjwt
psycopg2-binary
email-validator
python-jose[cryptography]
python-multipart